+++ PSP Project/Include/GameScreen.h
@@ -100,7 +100,7 @@ public:
 
 	virtual void draw();
 
-	virtual void update(u64 timePassed);
+	virtual void update(u32 timePassed);
 
 	const GameState GetGameState();
 	void SetGameState(const GameState &newState);
