+++ PSP Project/Include/IScreen.h
@@ -14,6 +14,6 @@ public:
 
 	virtual void draw() = 0;
 
-	virtual void update(u64 timePassed) = 0;
+	virtual void update(u32 timePassed) = 0;
 };
 #endif
